Add files via upload
This code calculates the Histogram of Oriented Gradient of the given images and uses it to calculate the cosine similarity of the images on the scale on 0 ( not similar) to 1 ( Similar images). On top of that this code also applies Sliding Window technique to crop the similar object in two images.